code/requirements.txt
matplotlib==3.7.3
numpy==1.25.2
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==9.5.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
